Refactor text rank into separate file
